shops working

+++ events/stringSelectMenu.js
@@ -1,7 +1,6 @@
 const player = require('../utils/playerUtils.js');
 const inv = require('../utils/inventoryUtils.js');
 const selector = require('../utils/messageTemplateUtils.js');
-const inventory = require('../utils/inventoryUtils.js');
 const { EmbedBuilder } = require('discord.js');
 const combat = require('../utils/combatUtils.js');
 const zoneData = require('../data/zones.json');
@@ -87,20 +86,46 @@ module.exports = {
                     return;
                 }
 
-                await interaction.message.delete();
+                selector.generateShopItemsSelector(interaction, interaction.values[0], "0", "0");
 
-                selector.generateShopItemSelector(interaction, interaction.values[0]);
+                await interaction.message.delete();
 
                 break;
             case 'shopItemChoice':
-
                 if(args[0] != interaction.user.id) {
                     interaction.channel.send("If you would like to shop with your own character, please use the t.shop commande yourself ! " + "<@" + interaction.user.id + ">");
                     return;
                 }
 
-                inventory.giveItem(interaction.user.id, interaction.values[0], 1);
+                if (interaction.message.components[2] == undefined) {
+                    var quantity = interaction.message.components[1].components[0].customId.split('-')[4];
+                    var shop = interaction.message.components[1].components[0].customId.split('-')[2];
+                } else {
+                    var quantity = interaction.message.components[2].components[0].customId.split('-')[4];
+                    var shop = interaction.message.components[2].components[0].customId.split('-')[2];
+                }
 
+                await interaction.message.delete();
+
+                selector.generateShopItemsSelector(interaction, shop, interaction.values[0], quantity);
+                break;
+            case 'shopAmountChoice':
+                if(args[0] != interaction.user.id) {
+                    interaction.channel.send("If you would like to shop with your own character, please use the t.shop commande yourself ! " + "<@" + interaction.user.id + ">");
+                    return;
+                }
+
+                if (interaction.message.components[2] == undefined) {
+                    var item = interaction.message.components[1].components[0].customId.split('-')[3];
+                    var shop = interaction.message.components[1].components[0].customId.split('-')[2];
+                } else {
+                    var item = interaction.message.components[2].components[0].customId.split('-')[3];
+                    var shop = interaction.message.components[2].components[0].customId.split('-')[2];
+                }
+
+                interaction.message.delete();
+
+                selector.generateShopItemsSelector(interaction, shop, item, interaction.values[0]);
                 break;
             default:
                 return;
